+++ src/main/java/cmc/controller/AuthController.java
@@ -1,11 +1,21 @@
 package cmc.controller;
 
+import cmc.common.ResponseCode;
+import cmc.common.ResponseDto;
 import cmc.service.AuthService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.security.Principal;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -15,5 +25,23 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
 
     private final AuthService authService;
+
+    @Operation(
+            summary = "회원 탈퇴",
+            description = "토큰에 해당하는 회원을 탈퇴합니다."
+    )
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "회원 탈퇴 성공"),
+            @ApiResponse(responseCode = "404", description = "User not found")
+    })
+    @DeleteMapping
+    public ResponseEntity<ResponseDto> deleteUser(Principal principal) {
+
+        Long tokenUserId = Long.parseLong(principal.getName());
+
+        authService.deleteUser(tokenUserId);
+
+        return ResponseEntity.status(HttpStatus.OK).body(new ResponseDto(ResponseCode.USER_DELETE_SUCCESS));
+    }
 
 }
